Added a method to grab every line on either an x or y axis. Used to find dangers to players and kill them on contact
(which is proving more difficult). Also added a distinction between speed and direction. Oh, and super powers.
Current problems include kill order (who should die first on a collision, determining whether the next should or not), moving things over to percent based systems (for varying screen types trying to collaborate on the same data) and plugging in gun to drive the system. Collision detection is what I forsee as the most complex problem.