Bug fixes, readability improvements, minor features
`claim.js` now accepts a number as an argument, making it compatible the `kill` event, allowing the game to be more responsive.
Brought in the newest version of gun which has number friendly path names and brought a fix for intermittent event firing, allowing me to simplify my `path`s and also took care of some pesky bugs.
`turn.js` is now able to determine whether it's own player was killed on the kill event without depending on players.me being a number, allowing the event to fire in more accurate circumstances.