Improve performance, created a waiting room, moved everything over to the new system, excluding collision detection. Now using webpack and a centralized server setup for waitroom and kicking logic.
Previously, everything was concatenated into one huge (ugly) file and
synchronized through an in-house event emitter. Now things are better
isolated (less global pollution) and can use node_modules without
horrible hacks.
It also allows sharing of server and client code.

Now everything is rendered straight from a synchronized gun cache
through the "synchronous" extension. Instead of trying to map everything to an
array, instead it just updates a local object representation of the gun
state and renders purely off of that, simplifying soooo much of the
logic.

Now, there is a concept of a waiting room, instead of the distributed
"everyone jump for the open spot as soon as you see it" paradigm. The
server elects the one who's been waiting longest as the new player. If
they don't respond, they are kicked from the
waiting room and the process is repeated.

If players are inactive for a period of time that places them off the
game board (latency compensated for), they are kicked from the game.
Also, if a client supporting the window.onbeforeunload event leaves the
game, it automatically removes the player.